Added wild card matching for pinned hosts
